# api/openapi.yaml
openapi: 3.0.3
info:
  title: ОбъектМаркет API
  version: 1.0.0
  description: 'API учебной системы закупки строительных материалов. Суммы — в копейках; реальные платежи и GPS отсутствуют. Демонстрационная доставка доступна только для строк адреса вида «Астрахань, улица, дом» или «г. Астрахань, улица, дом».'
servers:
  - url: /api/v1
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: om_session
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
...
    Product:
      type: object
      properties:
        id: {type: string, format: uuid}
        slug: {type: string}
        name: {type: string}
        category: {type: string}
        unit: {type: string}
        imageUrl: {type: string, nullable: true}
        description: {type: string}
        specs: {type: object, additionalProperties: true}
        priceFromKopecks: {type: integer, nullable: true}
...
  responses:
    Error:
      description: Ошибка с кодом и сообщением
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
paths:
...
  /products:
    get:
      summary: Поиск и фильтрация товаров
      parameters:
        - {name: q, in: query, schema: {type: string}}
        - {name: category, in: query, schema: {type: string}}
        - {name: sort, in: query, schema: {type: string, enum: [name, price_asc, price_desc]}}
        - {name: page, in: query, schema: {type: integer, minimum: 1}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
      responses:
        '200':
          description: Страница товаров
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/Product'}}
                  page: {type: integer}
                  total: {type: integer}
  /products/{id}:
    get:
      summary: Товар и предложения поставщиков
      parameters: [{name: id, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Карточка товара с offers}, '404': {$ref: '#/components/responses/Error'}}
  /supplier/offers:
    get:
      summary: Собственные предложения поставщика
      security: [{session: []}]
      responses: {'200': {description: Список предложений}}
    post:
      summary: Создать предложение
      security: [{session: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, warehouseId, priceKopecks, stock, deliveryDays, deliveryCostKopecks]
              properties:
                productId: {type: string, format: uuid}
                warehouseId: {type: string, format: uuid}
                priceKopecks: {type: integer, minimum: 1}
                stock: {type: integer, minimum: 0}
                deliveryDays: {type: integer, minimum: 0}
                deliveryCostKopecks: {type: integer, minimum: 0}
      responses: {'201': {description: Новое предложение}, '403': {$ref: '#/components/responses/Error'}}
  /supplier/offers/{id}:
    patch:
      summary: Изменить своё предложение
      security: [{session: []}]
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                priceKopecks: {type: integer, minimum: 1}
                stock: {type: integer, minimum: 0}
                deliveryDays: {type: integer, minimum: 0}
                deliveryCostKopecks: {type: integer, minimum: 0}
                active: {type: boolean}
      responses: {'200': {description: Изменённое предложение}, '404': {$ref: '#/components/responses/Error'}}
  /supplier/warehouses:
    get:
      summary: Склады текущего поставщика
      security: [{session: []}]
      responses: {'200': {description: Список складов}}
  /admin/categories:
    post:
      summary: Создать категорию, роль admin
      security: [{session: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: {type: string}
                slug: {type: string}
      responses: {'201': {description: Категория создана}}
  /admin/categories/{id}:
    patch:
      summary: Изменить категорию, роль admin
      security: [{session: []}]
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Категория изменена}}
  /admin/products:
    post:
      summary: Создать товар, роль admin
      security: [{session: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categoryId, slug, name, unit]
              properties:
                categoryId: {type: string, format: uuid}
                slug: {type: string}
                name: {type: string}
                unit: {type: string}
                imageUrl: {type: string, nullable: true}
                description: {type: string}
                specs: {type: object, additionalProperties: true}
      responses: {'201': {description: Товар создан}}
  /admin/products/{id}:
    patch:
      summary: Изменить товар, роль admin
      security: [{session: []}]
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Товар изменён}}
...
  /dispatch/drivers:
    get:
      summary: Доступные водители для назначения
      security: [{session: []}]
      responses: {'200': {description: 'Массив объектов {id, name}'}}
...
  /reports/summary:
    get:
      summary: Сводка администратора по учебным данным
      security: [{session: []}]
      responses: {'200': {description: Количество и суммы заказов, статусы поставок}}
...
  /geo/route:
    get:
      summary: Маршрут 2ГИС или схематический маршрут демо
      parameters:
        - {name: fromLon, in: query, required: true, schema: {type: number}}
        - {name: fromLat, in: query, required: true, schema: {type: number}}
        - {name: toLon, in: query, required: true, schema: {type: number}}
        - {name: toLat, in: query, required: true, schema: {type: number}}
      responses: {'200': {description: 'Источник, координаты [долгота, широта], дистанция и время'}, '502': {$ref: '#/components/responses/Error'}}